Move warnAboutDataLoss call into StoreWriter#processSelectionSet.
Since cache.modify calls cache.merge, and we don't want cache.modify ever
to trigger "Cache data may be lost..." warnings, it's more convenient if
the StoreWriter is responsible for calling warnAboutDataLoss, rather than
always performing that check in cache.merge.
