+++ src/cache/inmemory/writeToStore.ts
@@ -1,5 +1,5 @@
 import { SelectionSetNode, FieldNode, DocumentNode } from 'graphql';
-import { InvariantError } from 'ts-invariant';
+import { invariant, InvariantError } from 'ts-invariant';
 
 import {
   createFragmentMap,
@@ -29,7 +29,7 @@ import { cloneDeep } from '../../utilities/common/cloneDeep';
 
 import { ReadMergeContext } from './policies';
 import { NormalizedCache } from './types';
-import { makeProcessedFieldsMerger, FieldValueToBeMerged } from './helpers';
+import { makeProcessedFieldsMerger, FieldValueToBeMerged, fieldNameFromStoreName } from './helpers';
 import { StoreReader } from './readFromStore';
 import { InMemoryCache } from './inMemoryCache';
 
@@ -271,6 +271,22 @@ export class StoreWriter {
         mergedFields = policies.applyMerges(entityRef, mergedFields, context);
       }
 
+      if (process.env.NODE_ENV !== "production") {
+        Object.keys(mergedFields).forEach(storeFieldName => {
+          const fieldName = fieldNameFromStoreName(storeFieldName);
+          // If a merge function was defined for this field, trust that it
+          // did the right thing about (not) clobbering data.
+          if (!policies.hasMergeFunction(typename, fieldName)) {
+            warnAboutDataLoss(
+              entityRef,
+              mergedFields,
+              storeFieldName,
+              context.store,
+            );
+          }
+        });
+      }
+
       context.store.merge(dataId, mergedFields);
 
       return entityRef;
@@ -303,4 +319,82 @@ export class StoreWriter {
       out,
     });
   }
+}
+
+const warnings = new Set<string>();
+
+// Note that this function is unused in production, and thus should be
+// pruned by any well-configured minifier.
+function warnAboutDataLoss(
+  existingRef: Reference,
+  incomingObj: StoreObject,
+  storeFieldName: string,
+  store: NormalizedCache,
+) {
+  const getChild = (objOrRef: StoreObject | Reference): StoreObject | false => {
+    const child = store.getFieldValue<StoreObject>(objOrRef, storeFieldName);
+    return typeof child === "object" && child;
+  };
+
+  const existing = getChild(existingRef);
+  if (!existing) return;
+
+  const incoming = getChild(incomingObj);
+  if (!incoming) return;
+
+  // It's always safe to replace a reference, since it refers to data
+  // safely stored elsewhere.
+  if (isReference(existing)) return;
+
+  // If we're replacing every key of the existing object, then the
+  // existing data would be overwritten even if the objects were
+  // normalized, so warning would not be helpful here.
+  if (Object.keys(existing).every(
+    key => store.getFieldValue(incoming, key) !== void 0)) {
+    return;
+  }
+
+  const parentType =
+    store.getFieldValue<string>(existingRef, "__typename") ||
+    store.getFieldValue<string>(incomingObj, "__typename");
+  const fieldName = fieldNameFromStoreName(storeFieldName);
+  const typeDotName = `${parentType}.${fieldName}`;
+  // Avoid warning more than once for the same type and field name.
+  if (warnings.has(typeDotName)) return;
+  warnings.add(typeDotName);
+
+  const childTypenames: string[] = [];
+  // Arrays do not have __typename fields, and always need a custom merge
+  // function, even if their elements are normalized entities.
+  if (!Array.isArray(existing) &&
+      !Array.isArray(incoming)) {
+    [existing, incoming].forEach(child => {
+      const typename = store.getFieldValue(child, "__typename");
+      if (typeof typename === "string" &&
+          !childTypenames.includes(typename)) {
+        childTypenames.push(typename);
+      }
+    });
+  }
+
+  invariant.warn(
+`Cache data may be lost when replacing the ${fieldName} field of a ${parentType} object.
+
+To address this problem (which is not a bug in Apollo Client), ${
+  childTypenames.length
+    ? "either ensure all objects of type " +
+        childTypenames.join(" and ") + " have IDs, or "
+    : ""
+}define a custom merge function for the ${
+  typeDotName
+} field, so InMemoryCache can safely merge these objects:
+
+  existing: ${JSON.stringify(existing).slice(0, 1000)}
+  incoming: ${JSON.stringify(incoming).slice(0, 1000)}
+
+For more information about these options, please refer to the documentation:
+
+  * Ensuring entity objects have IDs: https://go.apollo.dev/c/generating-unique-identifiers
+  * Defining custom merge functions: https://go.apollo.dev/c/merging-non-normalized-objects
+`);
 }
